+++ routes/api/contacts-router.js
@@ -1,91 +1,22 @@
 import express from "express";
-import contactsServise from "../../models/contacts.js";
-import { HttpError } from "../../helpers/index.js";
-import Joi from "joi";
+import {
+  contactsGetController,
+  contactsGetByIdController,
+  contactsPostController,
+  contactsDeleteController,
+  contactsPutController,
+} from "../../controllers/contacts-controller.js";
 
 const contactsRouter = express.Router();
 
-const contactsAddSchema = Joi.object({
-  name: Joi.string().required().messages({
-    "any.required": `missing required name field`,
-  }),
-  email: Joi.string().required().messages({
-    "any.required": `missing required email field`,
-  }),
-  phone: Joi.string().required().messages({
-    "any.required": `missing required phone field`,
-  }),
-});
+contactsRouter.get("/", contactsGetController);
 
-contactsRouter.get("/", async (req, res, next) => {
-  try {
-    const result = await contactsServise.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.get("/:id", contactsGetByIdController);
 
-contactsRouter.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsServise.getContactById(id);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.post("/", contactsPostController);
 
-contactsRouter.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contactsServise.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.delete("/:id", contactsDeleteController);
 
-contactsRouter.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsServise.removeContact(id);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
-
-contactsRouter.put("/:id", async (req, res, next) => {
-  try {
-    const body = JSON.stringify(req.body);
-
-    if (body === "{}") {
-      console.log("no body!!");
-      throw HttpError(400, "missing fields");
-    }
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { id } = req.params;
-    const result = await contactsServise.updateContactById(id, req.body);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.put("/:id", contactsPutController);
 
 export default contactsRouter;
